Move MockUser class to pnc-mock module
